requirements.txt
fastapi
uvicorn
openai
langchain==0.1.16
langchain-openai==0.1.4
python-dotenv
faiss-cpu
tiktoken
openai==1.30.1
streamlit==1.33.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.2.0
langchain-openai==0.1.0
langchain-community==0.2.0
faiss-cpu==1.7.4
python-docx>=0.8.11
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.0.0